Update acquire_bar_controller
Acquire bar can work now.

+++ src/view/main_window_content/tabs/channels/stack_timepoint_settings.py
@@ -47,9 +47,9 @@ class stack_timepoint_frame(ttk.Frame):
         # Timepoints Label, spinbox defaults to 1.
         stack_timepoint_label_frame.filterwheel_label = ttk.Label(stack_timepoint_label_frame, text='Timepoints')
         stack_timepoint_label_frame.filterwheel_label.grid(row=1, column=label_position, sticky=(NSEW))
-        stack_timepoint_label_frame.exp_time_spinval = DoubleVar()
+        stack_timepoint_label_frame.exp_time_spinval = IntVar()
         stack_timepoint_label_frame.exp_time_spinbox = ttk.Spinbox(
-            stack_timepoint_label_frame, from_=0, to=5000.0, textvariable=stack_timepoint_label_frame.exp_time_spinval,
+            stack_timepoint_label_frame, from_=0, to=5000, textvariable=stack_timepoint_label_frame.exp_time_spinval,
             increment=1,
             width=9)
         stack_timepoint_label_frame.exp_time_spinbox.grid(row=1, column=input_position, sticky=(NSEW))
